+++ kit/dapp/src/components/blocks/xvp/approval-status-badge.tsx
